refactor(wamuseum-client): :memo: modified libs path
