+++ apps/wamuseum-client/libs/index.ts
@@ -1,4 +1,5 @@
 export * from './apollo-hydration-boundary'
 export * from './constants'
 export * from './create-apollo-client'
+export * from './registry'
 export * from './use-apollo'
