+++ apps/wamuseum-client/libs/registry/RegistryProvider.tsx
