+++ apps/wamuseum-client/libs/registry/StyledComponentsRegistry.tsx
