+++ apps/wamuseum-client/libs/registry/apollo-provider-registry/apollo-provider-registry.tsx
@@ -2,7 +2,7 @@ import { ApolloProvider } from '@apollo/client'
 import { useApollo } from 'libs'
 import { PropsWithChildren } from 'react'
 
-export default function ApolloProviderRegistry({ children }: PropsWithChildren) {
+export function ApolloProviderRegistry({ children }: PropsWithChildren) {
   const client = useApollo({})
   return <ApolloProvider client={client}>{children}</ApolloProvider>
 }
