+++ apps/wamuseum-client/libs/registry/index.ts
@@ -1,0 +1,1 @@
+export * from './apollo-provider-registry'
